move Formula class to new utils package
